+++ back/index.js
@@ -14,7 +14,7 @@ require('./launcher');
 
 if (process.platform === "win32") {
     const tempPath = path.join(process.env.LOCALAPPDATA, 'Temp');
-    fs.readdirSync(tempPath).filter(f => f.includes("frida")).forEach(f => fs.rmSync(tempPath + "\\" + f, {
+    fs.existsSync(tempPath) && fs.readdirSync(tempPath).filter(f => f.includes("frida")).forEach(f => fs.rmSync(tempPath + "\\" + f, {
         recursive: true,
         force: true
     }));
